Fix bug in fb_write_cell
The example code for fb_write_cell incorrectly writes the character at framebuffer[i] and the formatting options to framebuffer[i + 1]. This leads to problems with repeated calls to fb_write_cell, e.g. a for loop inside of printf.

Repro:
fb_write_cell(0, 'A', 0, 1);
fb_write_cell(1, 'B', 0, 1);
fb_write_cell(2, 'C', 0, 1);

The second call would write 'B' over the formatting options for the character 'A' (that is, framebuffer[1]). In actuality, 'B' should be written to framebuffer[2], with B's formatting options in framebuffer[3].