Remove colon after section .text in loader.s
This causes an issue later on in section 4 when trying to create a driver for the frame buffer. Specifically, removing the colon fixes a problem with passing string in function calls. Refer to: http://stackoverflow.com/questions/28709256/error-13-invalid-or-unsupported-executable-while-booting-simple-kernel-in-grub . Not sure why, but this fixed it for me when I had the same issue.